ingestor/requirements.txt
requests
kafka-python
python-dotenv
openaq